backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.9.2
httpx==0.27.2
numpy==1.26.4
Pillow==10.4.0
ImageHash==4.3.1
google-cloud-aiplatform==1.71.0
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-cloud-kms==3.1.0
google-generativeai==0.8.3
